[Server][Client] 서버권위 + 클라 예측 이동 + 스냅샷 보간 + 서버시간 동기화 + 핑측정 구현 완료
- [Client] JitterMeasurer 클래스 수정. (지터값이 너무 튀지않도록 제한 코드 작성)
- [Client] MyPlayer 클래스 수정. (멤버 변수 이름 수정, 장부 기록 함수 추가)
- [Client] NetworkClockManager 클래스 수정. (클라 처음 실행 될 때, 서버 시간 보간하지 않고 바로 대입)
- [Client] Player 클래스 수정. (클라와 서버 시간 동기화 작업 이후, OpponentMoveSyncByInterpolation 함수 개선)
- [Client] ServerPacketHandler 수정. (잘 못 된 값으로 지터 측정 부분 수정. g_client_total_time)
- [Client] g_total_client_time 전역 변수 추가. (클라이언트 실행 시간을 알기 위해)
- [Server] g_server_total_time 전역 변수 자료형을 float -> double 수정.
- [Server] Player 클래스 수정. (Player의 timestamp 값에 g_total_server_time 값으로 대입)
- [Server][Client] 보간용 Render Buffer 사이즈 60, 클라이언트 입력 장부 120으로 수정

+++ Client/Client.cpp
@@ -23,7 +23,7 @@ BOOL                InitInstance(HINSTANCE, int);
 LRESULT CALLBACK    WndProc(HWND, UINT, WPARAM, LPARAM);
 INT_PTR CALLBACK    About(HWND, UINT, WPARAM, LPARAM);
 
-float g_client_total_time = 0.0f;
+double g_client_total_time = 0.0f;
 
 HWND ghWnd;
 CGameFramework gGameFramework;
